refactor(api): split period controllers into 'core' and 'assignment' for clearer code organization
